requirement.txt
torch==2.0.1
streamlit==1.22.0
transformers==4.29.2
requests
pillow
